+++ src/pages/member/[id].tsx
@@ -1,12 +1,11 @@
-import { MemList } from "../../../types/member";
 import { client } from "../../../libs/client";
 import Link from "next/link";
 import styles from "@/styles/sections/member/Member.module.scss";
 import Hero from "@/components/sections/hero/Hero";
+import Member from "@/components/sections/member/Member";
+import type { GetStaticProps, InferGetStaticPropsType, NextPage } from "next";
 
-type Props = {
-  member: MemList;
-};
+type Props = InferGetStaticPropsType<typeof getStaticProps>;
 
 export const getStaticPaths = async () => {
   const data = await client.get({ endpoint: "member" });
@@ -20,7 +19,7 @@ export const getStaticPaths = async () => {
   };
 };
 
-export const getStaticProps = async (context: any) => {
+export const getStaticProps: GetStaticProps = async (context: any) => {
   const id = context.params.id;
   const data = await client.get({ endpoint: "member", contentId: id });
 
@@ -30,39 +29,15 @@ export const getStaticProps = async (context: any) => {
     },
   };
 };
-const MemberList = (member: Props) => {
-  const body = member.member.body;
+const MemberList: NextPage<Props> = ({ member }) => {
   return (
     <section className={styles.section}>
       <Hero />
-      <div className={styles.container}>
-        {body.map((member: any, index: number) => (
-          <div key={index} className={styles.card}>
-            <div className={styles.item}>
-              <img src={member.img.url} className={styles.img} alt="memimg" />
-              <div className={styles.content}>
-                <h1>{member.name}</h1>
-                <p>
-                  <span>好きなよさこいソング</span>
-                  <br />
-                  {member.song}
-                  <br />
-                  <span>休日の過ごし方</span>
-                  <br />
-                  {member.holiday}
-                  <br />
-                  <span>好きな有名人</span>
-                  <br />
-                  {member.celebrity}
-                  <br />
-                  <span>ひとこと</span>
-                  <br />
-                  {member.comment}
-                </p>
-              </div>
-            </div>
-          </div>
-        ))}
+      <Member member={member} />
+      <div className={styles.btn}>
+        <Link href="/member">
+          <button>一覧へ戻る</button>
+        </Link>
       </div>
     </section>
   );
