+++ src/styles/sections/member/Member.module.scss
@@ -74,3 +74,20 @@
 .card:hover img {
   transform: scale(1.5);
 }
+
+.btn {
+  margin: 0 auto;
+  margin-top: 40px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 200px;
+  height: 50px;
+  background-color: $bk;
+  color: $wh;
+  font-size: 16px;
+  font-weight: bold;
+  border-radius: 5px;
+  cursor: pointer;
+  transition: 0.3s;
+}
